Two bugs and sticker support.
Telegram users can block the bot, and an attempt to send a message to
such user end up with an error. But not anymore. There's new 'blocked'
table for users, which is blocking the bot, and transitions between
Contacts and Blocked are automatic.
And users can change there names/usernames. Now they won't be
duplicated in the database.
Stickers are just for fun.